dom/base/test/test_bug1384658.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1384658
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1384658</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1384658 **/

  SimpleTest.waitForExplicitFinish();
  var win = window.open("http://example.com/tests/dom/base/test/window_bug1384658.html",
                        "_blank", "width=100,height=100");
  var bc = new BroadcastChannel("test_channel");
  bc.onmessage = ev => {
    ok(ev.data.success, "We didn't hang");
    bc.close();
    win.close();
    SimpleTest.finish();
  };

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1384658">Mozilla Bug 1384658</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
